requirements.txt
asgiref==3.9.1
black==25.1.0
cfgv==3.4.0
click==8.2.1
distlib==0.4.0
Django==4.2.7
django-extensions==4.1
filelock==3.19.1
flake8==7.3.0
identify==2.6.13
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pre_commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-decouple==3.8
PyYAML==6.0.2
sqlparse==0.5.3
virtualenv==20.34.0
# Add these for Google OAuth
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
requests==2.31.0

echo "gunicorn==21.2.0" >> requirements.txt
echo "whitenoise==6.6.0" >> requirements.txt
echo "psycopg2-binary==2.9.9" >> requirements.txt